feat: 지하철 노선 목록 조회 기능 추가
- 추가로 기존 StationAcceptanceTest 에 작성하던 노선 관련 테스트 코드를 별도 클래스(LineAcceptanceTest)로 분리
