+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -1,9 +1,6 @@
 package nextstep.subway.station;
 
 import io.restassured.RestAssured;
-import io.restassured.path.json.JsonPath;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import io.restassured.response.ValidatableResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -117,88 +114,6 @@ public class StationAcceptanceTest {
                 .jsonPath()
                 .getList("name", String.class);
         assertThat(stationNames).isEmpty();
-    }
-
-    /*
-     * When 지하철 노선을 생성하면
-     * Then 지하철 노선 목록 조회 시 생성한 노선을 찾을 수 있다
-     * */
-    @DisplayName("지하철 노선을 생성한다.")
-    @Test
-    void createLine() {
-        createStation("강남역");
-        createStation("양재역");
-        List<Long> stationsIds = fetchStation()
-                .extract()
-                .jsonPath()
-                .getList("id", Long.class);
-
-        // When
-        ExtractableResponse<Response> response = createLine("신분당선", "bg-red-600", stationsIds.get(0), stationsIds.get(1), 10).extract();
-
-        // Then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        JsonPath responseBody = response.jsonPath();
-        System.out.println(responseBody.toString());
-    }
-
-    /*
-     * Given 2개의 지하철 노선을 생성하고
-     * When 지하철 노선 목록을 조회하면
-     * Then 지하철 노선 목록 조회 시 2개의 노선을 조회할 수 있다.
-     * */
-    @DisplayName("지하철 노선 목록을 조회한다.")
-    @Test
-    void getLines() {
-
-    }
-
-    /*
-     * Given 지하철 노선을 생성하고
-     * When 생성한 지하철 노선을 조회하면
-     * Then 생성한 지하철 노선의 정보를 응답받을 수 있다.
-     * */
-    @DisplayName("지하철 노선을 조회한다.")
-    @Test
-    void getLine() {
-
-    }
-
-    /*
-    * Given 지하철 노선을 생성하고
-    * When 생성한 지하철 노선을 수정하면
-    * Then 해당 지하철 노선 정보는 수정된다
-    * */
-    @DisplayName("지하철 노선을 수정한다.")
-    @Test
-    void updateLine() {
-
-    }
-
-    /*
-    * Given 지하철 노선을 생성하고
-    * When 생성한 지하철 노선을 삭제하면
-    * Then 해당 지하철 노선 정보는 삭제된다
-    * */
-    @DisplayName("지하철 노선을 삭제한다.")
-    @Test
-    void deleteLine() {
-
-    }
-
-    private ValidatableResponse createLine(String name, String color, long upStationId, long downStationId, int distance) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("name", name);
-        params.put("color", color);
-        params.put("upStationId", upStationId);
-        params.put("downStationId", downStationId);
-        params.put("distance", distance);
-
-        return RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
-                .then().log().all();
     }
 
     private ValidatableResponse createStation(String name) {
